requirements.txt
# Core
python-dotenv>=1.0.0
pydantic>=2.0.0
requests>=2.31.0

# AI / LLM
groq>=0.4.0
openai>=1.10.0
google-generativeai>=0.3.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0

# Vector DB / Memory (Pin to avoid Pydantic issues)
chromadb>=0.4.22
numpy<2.0.0

# Tools
beautifulsoup4>=4.12.0
mss>=9.0.1
pyautogui>=0.9.54
clipboard>=0.0.4
psutil>=5.9.0
screen-brightness-control>=0.8.0
pymupdf>=1.23.0
docker>=7.0.0
fpdf2>=2.7.0
python-pptx>=0.6.21

# Integrations
python-telegram-bot>=20.0
youtube-search-python>=1.6.6
duckduckgo-search>=6.0.0
playwright>=1.40.0
watchdog>=3.0.0
pathspec>=0.11.0
pathspec>=0.11.0
rich>=13.0.0
win11toast>=0.33
sounddevice>=0.4.6
scipy>=1.11.0
openai-whisper>=20231117

# Distribution
build
twine
twine
wheel
setuptools
cryptography>=42.0.0
